WORKDONE: Cleaned up HTML Framework, CSS preperation
