+++ src/Components/Dashboard.js
@@ -4,12 +4,12 @@ import { Link } from 'react-router-dom';
 export default class Dashboard extends Component {
   render() {
     return (
-      <div>
-        <h1>Dashboard</h1>
+      <div className='main'>
+        <h2>Dashboard</h2>
         <p>
           Welcome, {this.props.user.name}!
         </p>
-        <Link to='/settings'>Settings</Link>
+        <Link to='/settings'><button>Settings</button></Link>
       </div>
     )
   }
