+++ src/Components/Header.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 export default class Header extends Component {
   render() {
     return (
-      <div>
+      <div className='header'>
         <h1>Affirbisiaks</h1>
       </div>
     )
