+++ src/Components/Login.js
@@ -1,10 +1,22 @@
 import React, { Component } from 'react'
+import GoogleLogin from 'react-google-login'
 
-export default class Login extends Component {
+const responseGoogle = (response) => {
+  console.log(response);
+}
+
+export default class Main extends Component {
   render() {
     return (
-      <div>
-        league of losers
+      <div className='main'>
+        <p>Login Using Google</p>
+        <GoogleLogin
+          clientId={process.env.REACT_APP_CLIENT_ID}
+          buttonText="Login"
+          onSuccess={ (token) => {this.props.event(token.profileObj)} }
+          onFailure={responseGoogle}
+          cookiePolicy={'single_host_origin'}
+        />
       </div>
     )
   }
