+++ src/utils/update.js
@@ -1,7 +1,6 @@
 const url = 'https://affirbisiaks.herokuapp.com';
 
 export async function update(userInfo) {
-    console.log(userInfo);
     const data = await fetch(`${url}/api/v1/users/${userInfo.googleId}`, {
         method: 'PATCH',
         headers: {
